Gallery2: put a itemPath to the bundle when start photopage.
In some cases,  it request to start photopage, but
the deletions is not cleared, it get 0 when get
media item count, so it can not get a itemPath to
create a photodata adapter when create photopage,
then it request to finish state when resume, then a
NullPointerException happens.

We put a itemPath to the bundle when start photopage.
So it can create a photodata adapter, and do not finish
the photopage state.

CRs-Fixed: 521735

Change-Id: I539bcfa35db35fdf0fe9004470e7a77acc65a94d